Add a separate command with a separate permission to leave a clan
